# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Zenith API Documentation",
        "description": "Documentation for Zenith API endpoints",
        "contact": {
            "email": "admin@zenith.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/reviews": {
            "post": {
                "tags": [
                    "Reviews"
                ],
                "summary": "Create a review",
                "description": "Store a newly created resource in storage.",
                "operationId": "f9bf17f06508aa2dd6aecbb9771092da",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "id_produk",
                                    "id_pesanan",
                                    "rating",
                                    "komentar"
                                ],
                                "properties": {
                                    "id_produk": {
                                        "type": "integer"
                                    },
                                    "id_variant": {
                                        "type": "integer"
                                    },
                                    "id_pesanan": {
                                        "type": "integer"
                                    },
                                    "rating": {
                                        "type": "integer",
                                        "maximum": 5,
                                        "minimum": 1
                                    },
                                    "komentar": {
                                        "type": "string"
                                    },
                                    "images[]": {
                                        "type": "array",
                                        "items": {
                                            "type": "string",
                                            "format": "binary"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Review created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Already reviewed"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/reviews/{productId}": {
            "get": {
                "tags": [
                    "Reviews"
                ],
                "summary": "Get reviews for a product",
                "operationId": "f65e3186e70d8557104cb09c3cc67fb0",
                "parameters": [
                    {
                        "name": "productId",
                        "in": "path",
                        "description": "Product ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of reviews",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}